requirements.txt
Django==4.1
psycopg2-binary==2.9.9
gunicorn==21.2.0

djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.4
 django_crispy_forms==1.9.2